// miniprogram/typings/entity/carProduct.ts
export interface ICarData {
	id: string;
	name: string;
	displacement: string;
	specifications: string;
	energy: string;
	volume: number;
	actuation: string;
	status: boolean;
	firstTag: string;
	secondTag: string;
	thirdTag: string;
	fourthTag: string;
	originalPrice: number;
	currentPrice: number;
	carImage: string;
	fuelOilNumber: string;
	seats: number;
	brand: string
}